My Captain Data Analytics with Python task-4
My Captain Data Analytics with Python task-4 : Record your productivity level(1-5) 5 being most productive and 1 being least productive and the time of the day(From the time you wake up to when you sleep). The expected result is that the time of the day affects your productivity. Perform a bivariate analysis by plotting a bar graph where x axis will be the time of the day and y axis will be the